Command requires validation; Moved Lamar methods to separate classes
