+++ lamar/test/Commands/LamarCommandExtensionsTests.cs
@@ -3,11 +3,9 @@
 // Licensed under the Apache License, Version 2.0
 // See https://opensource.org/licenses/Apache-2.0 or the LICENSE file in the repository root for the full text of the license.
 
-using FluentValidation;
 using Lamar;
 using Lamar.Scanning.Conventions;
 using Moq;
-using Shipwright.Commands;
 using Shipwright.Commands.Internal;
 using Shipwright.Validation;
 using System;
@@ -15,72 +13,13 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Shipwright.Core.Lamar.Test
+namespace Shipwright.Commands
 {
-    public class CoreLamarExtensionsTests
+    public class LamarCommandExtensionsTests
     {
         private ServiceRegistry registry = new ServiceRegistry();
 
-        public class AddValidationAdapter : CoreLamarExtensionsTests
-        {
-            private ServiceRegistry method() => registry.AddValidationAdapter();
-
-            [Fact]
-            public void requires_registry()
-            {
-                registry = null!;
-                Assert.Throws<ArgumentNullException>( nameof( registry ), method );
-            }
-
-            [Fact]
-            public void adds_validation_adapter()
-            {
-                var actual = method();
-                Assert.Same( registry, actual );
-
-                using var container = new Container( registry );
-                var instance = container.GetInstance<IValidationAdapter<AddValidationAdapter>>();
-                Assert.IsType<ValidationAdapter<AddValidationAdapter>>( instance );
-            }
-        }
-
-        public class AddValidators : CoreLamarExtensionsTests
-        {
-            private IAssemblyScanner scanner;
-
-            private IAssemblyScanner method() => scanner.AddValidators();
-
-            public class FakeValidator : AbstractValidator<AddValidators> { }
-
-            [Fact]
-            public void requires_scanner()
-            {
-                scanner = null!;
-                Assert.Throws<ArgumentNullException>( nameof( scanner ), method );
-            }
-
-            [Fact]
-            public void adds_discovered_validator_as_singleton()
-            {
-                registry.Scan( scanner =>
-                {
-                    this.scanner = scanner;
-                    scanner.AssemblyContainingType<AddValidators>();
-
-                    var actual = method();
-                    Assert.Same( this.scanner, actual );
-                } );
-
-                using var container = new Container( registry );
-                var instance = container.GetInstance<IValidator<AddValidators>>();
-                Assert.IsType<FakeValidator>( instance );
-
-                var second = container.GetInstance<IValidator<AddValidators>>();
-                Assert.Same( instance, second );
-            }
-        }
-
-        public class AddCommandDispatch : CoreLamarExtensionsTests
+        public class AddCommandDispatch : LamarCommandExtensionsTests
         {
             private ServiceRegistry method() => registry.AddCommandDispatch();
 
@@ -103,10 +42,9 @@ namespace Shipwright.Core.Lamar.Test
             }
         }
 
-        public class AddCommandHandlers : CoreLamarExtensionsTests
+        public class AddCommandHandlers : LamarCommandExtensionsTests
         {
             private IAssemblyScanner scanner;
-
             private IAssemblyScanner method() => scanner.AddCommandHandlers();
 
             public class FakeCommandHandler : CommandHandler<FakeGuidCommand, Guid>
@@ -141,7 +79,7 @@ namespace Shipwright.Core.Lamar.Test
             }
         }
 
-        public class AddCommandValidation : CoreLamarExtensionsTests
+        public class AddCommandValidation : LamarCommandExtensionsTests
         {
             private ServiceRegistry method() => registry.AddCommandValidation();
 
@@ -171,7 +109,7 @@ namespace Shipwright.Core.Lamar.Test
             }
         }
 
-        public class AddCommandCancellation : CoreLamarExtensionsTests
+        public class AddCommandCancellation : LamarCommandExtensionsTests
         {
             private ServiceRegistry method() => registry.AddCommandCancellation();
