+++ src/Commands/Command.cs
@@ -3,6 +3,7 @@
 // Licensed under the Apache License, Version 2.0
 // See https://opensource.org/licenses/Apache-2.0 or the LICENSE file in the repository root for the full text of the license.
 
+using Shipwright.Validation;
 using System;
 
 namespace Shipwright.Commands
@@ -12,7 +13,7 @@ namespace Shipwright.Commands
     /// </summary>
     /// <typeparam name="TResult">Type of the command result.</typeparam>
 
-    public abstract record Command<TResult> { }
+    public abstract record Command<TResult> : IRequiresValidation { }
 
     /// <summary>
     /// Defines a command that does not return a result when executed.
